for the first time sync subject might not be present so detele the child entites only if subject is there

+++ packages/openchs-android/src/service/SubjectMigrationService.js
@@ -120,6 +120,7 @@ class SubjectMigrationService extends BaseService {
 
     removeEntitiesFor({subjectUUID}) {
         const subject = this.entityService.findByUUID(subjectUUID, Individual.schema.name);
+        if (_.isNil(subject)) return;
         const db = this.db;
         db.write(() => {
             db.delete(subject.encounters);
